api/pyproject.toml
[tool.poetry]
name = "hrsafindings-api"
version = "0.1.0"
description = ""
authors = ["Yevgeny Bulochnik <yevgeny.bulochnik@gmail.com>"]
packages = [
    { include = "app" },
    { include = "app/**/*.py" }
]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask-sqlalchemy = "^2.4.1"
flask-marshmallow = "^0.12.0"
python-dotenv = "^0.13.0"
click = "^7.1.2"
marshmallow-sqlalchemy = "^0.23.0"
psycopg2-binary = "^2.8.5"
webargs = "^6.1.0"
flask-migrate = "^2.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
ipython = "^7.14.0"
pyyaml = "^5.3.1"

[tool.poetry.plugins."flask.commands"]
"seed" = "app.cli:seed"
"ipython" = "app.cli:ipython"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
